requirements.txt
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
python-decouple==3.8
drf-spectacular==0.28.0
